Remove some low-hanging duplication beween GenJVM / GenASM.
Refactoring of the the zero-risk variety. The extraction of
`ExcludedForwarderFlags` is the motivation here; we'll need that
next commit to fix SI-5894.
